🤖 Implement PromptChatCommand for chat functionality
Add a new PromptChatCommand class to handle chat functionality with the
prompt bot. This class includes methods for generating the command,
running the chat, and reading multi-line input. The main Program.cs file
has been updated to use the new PromptChatCommand class, simplifying the
code and improving organization.
